Changes: - Restaurant id is now Long type - User tenantId is now renamed to organizationId - updated relevant code - updated relevant tests

+++ src/main/java/com/hackybear/hungry_scan_core/controller/cms/RestaurantController.java
@@ -34,7 +34,7 @@ public class RestaurantController {
 
     @PostMapping("/show")
     @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<Map<String, Object>> show(@RequestBody Integer id) {
+    public ResponseEntity<Map<String, Object>> show(@RequestBody Long id) {
         return responseHelper.getResponseEntity(id, restaurantService::findById);
     }
 
@@ -53,7 +53,7 @@ public class RestaurantController {
 
     @DeleteMapping("/delete")
     @PreAuthorize("hasAnyRole('MANAGER', 'ADMIN')")
-    public ResponseEntity<Map<String, Object>> delete(@RequestBody Integer id) {
+    public ResponseEntity<Map<String, Object>> delete(@RequestBody Long id) {
         return responseHelper.buildResponse(id, restaurantService::delete);
     }
